Move out tests from pytestqt package
