+++ tests/test_exceptions.py
